requirements.txt
skypilot[aws,gcp,kubernetes]
unique-names-generator
dill 
llama-index
llama-index-finetuning
fire
